app/requirements.txt
pandas==2.0.3
plotly==5.17.0
psycopg2-binary==2.9.7
streamlit==1.27.0
streamlit-autorefresh==1.0.1